==> v0/centralized/sensor_anomaly/2_data_analysis.ipynb <==
# Imports
import matplotlib.patches as mpatches
import matplotlib.pyplot as plt
import pandas as pd
import ydata_profiling
from sklearn.decomposition import PCA

# Functions

def plot_long_lat(df):
    # Plotting longitude and latitude
    plt.scatter(x=df['x'], y=df['y'])
    plt.xlabel("Latitude")
    plt.ylabel("Longitude")
    plt.show() # Plot show

    
    
def apply_pca(X, n_features):
    # Applying the PCA
    pca = PCA(n_components=n_features)
    X_reduced = pca.fit_transform(X)
    # Features DataFrame with the dimension reduction
    X_reduced = pd.DataFrame(data=X_reduced, columns=[f'feature_{i + 1}' for i in range(n_features)])
    #  Features DataFrame with the reduced dimensions and explained variance
    return X_reduced, pca.explained_variance_ratio_



def plot_3d(benign_reduced, jamming_reduced, spoofing_reduced):
    # Configuring the figure
    fig = plt.figure(figsize=(16,10))
    plt.rcParams.update({'font.size': 16})
    # Adding subplot
    ax = fig.add_subplot(111, projection='3d')
    # Plotting the samples of each class
    ax.plot(benign_reduced['feature_1'], benign_reduced['feature_2'], benign_reduced['feature_3'], '.', markersize=10, color='#2E8B57', alpha=0.8)
    ax.plot(jamming_reduced['feature_1'], jamming_reduced['feature_2'], jamming_reduced['feature_3'], '.', markersize=10, color='#B22222', alpha=0.8)
    ax.plot(spoofing_reduced['feature_1'], spoofing_reduced['feature_2'], spoofing_reduced['feature_3'], '.', markersize=10, color='#DB9046', alpha=0.8)
    # Setting the font and spacing for each axis
    ax.set_xlabel('Feature 1', labelpad=20, fontsize=20)
    ax.set_ylabel('Feature 2', labelpad=20, fontsize=20)
    ax.set_zlabel('Feature 3', labelpad=20, fontsize=20)
    # Graph legend
    benign = mpatches.Patch(color='#2E8B57', label='Benign')
    jamming = mpatches.Patch(color='#B22222', label='GPS Jamming')
    spoofing = mpatches.Patch(color='#DB9046', label='GPS Spoofing')
    ax.legend(handles=[benign, jamming, spoofing], markerscale=2, ncol=2, loc=1, title='Category')
    plt.show() # Plot show

# Main

# Dataset path
dataset_path = '/home/leandro/remy-project/centralized/datasets/UAVGPSAttacks/'

# Load dataset
uav_df = pd.read_csv(dataset_path + 'data.csv')
# Load normalized dataset
uav_df_norm = pd.read_csv(dataset_path + 'data_norm.csv')

# Variation of longitude and latitude - Benign
plot_long_lat(uav_df[uav_df['class'] == 0])

# Variation of longitude and latitude - GPS Jamming
plot_long_lat(uav_df[uav_df['class'] == 1])

# Variation of longitude and latitude - GPS Spoofing
plot_long_lat(uav_df[uav_df['class'] == 2])

# Applying PCA to data for each category
plot_3d(apply_pca(uav_df_norm[uav_df_norm['class'] == 0], 3)[0],
        apply_pca(uav_df_norm[uav_df_norm['class'] == 1], 3)[0],
        apply_pca(uav_df_norm[uav_df_norm['class'] == 2], 3)[0])

# Generate profile report
report = uav_df.profile_report(title="UAV Attack Dataset", html={"style": {"full_width": True}})
# Save report to file
report.to_file("profile_report.html")
